+++ CHANGELOG.md
@@ -11,9 +11,11 @@ Keep it human-readable, your future self will thank you!
 ## [Unreleased]
 
 ### Added
+- ci: changelog release updater
 - earthkit-data replaces climetlab
 
 ### Changed
+- ci: updated workflows on PR and releases to use reusable actions
 
 ### Removed
 - climetlab